energies_q15 ok

+++ lib/mel.h
@@ -16,8 +16,8 @@ void apply_filterbank(
 
 void create_filterbank_q15(int16_t filterbank[NUM_FILTERS][NFFT/2 + 1], int sample_rate);
 void apply_filterbank_q15(
-    int16_t power_spectrum_frame[NFFT/2 + 1],
-    int16_t filterbank[NUM_FILTERS][NFFT/2 + 1],
+    int32_t power_spectrum_frame[NFFT/2 + 1],
+    int32_t filterbank[NUM_FILTERS][NFFT/2 + 1],
     int16_t energies_q15[NUM_FILTERS]  // Saída em Q1.15 (log)
 );
